Expand test coverage and fix midring dtype bug
- Fix dtype mismatch in healpixl_ring_to_xy: midring() used .astype(int)
  (int64 on 64-bit systems) while largering() produced int32, causing
  jax.lax.cond to fail during jax.grad tracing. Changed to .astype(hp.dtype).

- Export ring_above and get_ring_info2 from __init__.py (they were public
  functions with docstrings but missing from the package API).

- Add 22 new tests across 5 new classes:
  * TestCoordinateConversions: unit norm, round-trip, and analytical value
    checks for ang2vec, ang2vec_radec, vec2ang, vec2ang_radec.
  * TestPixelConversionsExtended: vec2pix and ang2pix_radec vs healpy
    reference, pix2vec vs healpy.pix2vec, pix2ang dx/dy offset parameters.
  * TestInternalFunctions: get_ring_info2 against reference formula for all
    rings at nsides 2-32; ring_above monotonicity check.
  * TestSchemeConversions: scheme2bighpxy/bighpxy2scheme round-trips (ring
    and nest), get_patch centre-is-input, patch/neighbours consistency.
  * TestUtilities: nside2npix, npix2nside, get_nside.
  * TestDifferentiability: jax.grad through pix2vec and pix2ang_radec w.r.t.
    dx/dy; interpolation weight sum-to-1 and pixel index bounds.

- Cap test_neighbours at nside=128 to prevent OOM (nside=1024 vmaps over
  12.6M pixels).

+++ healjax/healjax.py
@@ -402,13 +402,13 @@ def healpixl_ring_to_xy(nside, hp):
         v = F1 * nside - ringind - 1
         h = 2 * postlongind - s - F2*nside
         h = h - R
-        prex = ((v + h) / 2).astype(int)
-        prey = ((v - h) / 2).astype(int)
+        prex = ((v + h) / 2).astype(hp.dtype)
+        prey = ((v - h) / 2).astype(hp.dtype)
 
         tweak = jnp.logical_or((v != (prex+prey)), (h != (prex-prey)))
         h = h + tweak
-        x = ((v + h) / 2).astype(int)
-        y = ((v - h) / 2).astype(int)
+        x = ((v + h) / 2).astype(hp.dtype)
+        y = ((v - h) / 2).astype(hp.dtype)
         return bighp, x, y
     
     def largering(): # otherwise
